resource: move commit status adapter functions to a dedicated file
